Add files via upload
TEFILLIN CHILDREN MINISTRY
Child Evangelism - Mission And Outreach To Children
Tefillin Children Ministry outreach focuses in reaching children with the good news of Jesus in Ethiopia.
We organize missions and outreach in places where children have not been reached and where the Gospel has not been communicated clearly.
This happens with partnership with churches and other ministries that we share the same passion in reaching children.
In our missions and outreaches we welcome groups and individuals who are passionate in reaching children and spreading the love of Jesus in any capacity they can.
Vision – Winning children to Jesus
OUR GOALS
•	Presenting Christ with clarity to every child.
•	Reaching the unreached children in the rural area and places where the Gospel has not been taken to.
•	Working and supporting the churches for effective children outreach and mission.
•	Raising missionaries for children in Ethiopia.
WHERE WE DO IT
•	In Schools
•	Homes
•	Sports grounds
•	Open air crusades
AND MORE
HOW WE DO IT THROUGH
1.	Dramas
2.	Poems
3.	Sports
4.	Mercy ministry
5.	Concerts
WHEN DO OUR OUTREACH AND MISSIONS HAPPEN
•	Missions– Most of our missions take place during school holidays whereby we are able to reach many children when they are out of school for holidays.
•	Our outreach in schools– happens weekly during classes whereby we are able to run a pastoral program with permission from the school, this has seen so many lives of children transformed and matured spiritually current we are operating in eight schools also we are able to hold one day outreach in long distance schools.
•	Concerts– Our concerts happen on weekends once in every month this takes place mostly on Sunday afternoon were by we are able to bring children together from different places and also invite their care takers this is a time of worship.
•	Mercy Ministry is a time we welcome people to give or donate what they can in order to meet the needs of many children who are going without food, clothes, education and so on. We welcome individuals, churches, organization to join and partner to reach our children. For more information on how you can be involved contact us.
CHALLENGES FACING OUTREACH TO CHILDREN IN ETHIOPIA
•	Lack of enough resources.
•	Lack of enough children workers.
•	Lack of vision in reaching children.
•	Lack of passion.
•	Lack of financial support for rural long distance missions.
•	Lack of our own mission bus which would make it easy for transportation.
Child Discipleship And Mentorship – Mentoring the next generation of leaders
As we reach children with the good news we are also making sure they are rooted in Christ through discipleship and mentorship for six month or more
Strategic objectives
•	To train young leaders who serve in super kids Africa! – give them a vision of what Christian leadership is about and allow them to reach the next generation
•	Use school pastoral programs as an opportunity for children discipleship and develop a kids club that is focused on kids mentorship
•	Strengthen families and help them discipleship their children
•	Passing the button to the next generation and the next leader programs
•	Unlocking the future destiny of the child seminars, where the parents will stand in the gap of their children.
Mentors
Mentors play an important role in the nurturing process of kids, so the church has to do its best to have good mentors. Mentors should be those who have a burden to serve children and at the same time be prepared for effective ministry through training and developing good characteristics. The number of mentors will be based on the age groups of children in the church.
Training mentors
Mentors will be trained well to enable them to teach the children in a Mature and Systematized way to ensure they are fruitful in their ministry. The training will be offered in a short course, seminars or in an institution. The quantity of
Mentors will be increased based on the number of children in the church
Venue: churches, schools, home 


Tefillin School Of Children Ministry -Equipping and empowering children workers for effective ministry
School of children ministry offers different programs to children ministers and organizations that work with children, from team building, short courses on children ministries and more.
The Program provides a 9 months certificate on Vibrant, growing, fruitful children ministry. The program prepares thoroughly the student or the children minister to serve in children ministry, para –church and any institution that serve with children.
Vision-raising effective children ministers
Mission-empower and equip children ministers to reach and disciple children of Africa through the whole council of God.
Goals
•	raising children missionaries for Africa
•	Equip children ministers for effective ministry
•	Mobilization of children worker in ministries and children program
•	Developing children curriculums for mentorship and Sunday school
•	empowering children ministers through trainings and exposure for experience on children
•	Children workers prayer net.



PROGRAMS
9-month certificate program
 This is a practical child ministers training that shapes the teacher and children ministry that is vibrant, growing and fruitful, the training takes place few days a month and practical application that is done through follow up.
Short courses and training 
We train Sunday school teachers or children organization on special lessons and topics of their choice depending on their need and the goal they want to achieve
Children ministers’ network
Tefillin Children Ministry is a Program that takes place every second Sunday of the month starting from 2:30 to 4pm. the program, allows children workers or ministers to be
refreshed, share testimonies, be empowered and added more skills on handling children and more and more find time to pray for children needs in their area of work.
Objectives for the Children ministers’ network
1.	Strengthening the bond between children workers/ministers in Ethiopia for effective ministry to children and share ideas
2.	Refreshing children ministers through different activities
3.	To share testimonies and challenges facing Children ministries.
4.	Monthly empowerment and addition of skills on handling children.
5.	Time of prayer and fellowship.
Tefillin School of children ministry ministers activities.
We organize children workers activities in order to build better leaders and unity and bond among the children workers
•	Team building
•	Retreats
•	Fun day
•	Teachers fellowship
•	Teachers view post and
•	Opinion
•	Teachers fun outing (trips, breakfast,, sports
•	Conferences and seminars
Children ministry consultancy  
We offer intensive and strategic children ministry consultancy on vibrant, growing, fruitful ministry to children, teens, and youth.

Our project
Building school of children ministry
It’s our desire to establish a school where children workers and ministers can be fully equipped and find mentorship   for effective work to children .The school will be able to give children workers and minister’s opportunities to grow their calling on children and expose them to different environments and ministries.
In the school we hope to establish a retreat center where children workers and ministers can find rest and time of spiritual refreshment.